Undo an unnecessary fix

+++ test/test_hover.jl
@@ -31,7 +31,7 @@ end
 end
 testmodule
 """
-LanguageServer.process(LanguageServer.JSONRPC.Request{Val{Symbol("textDocument/didOpen")},LanguageServer.DidOpenTextDocumentParams}(Int64(0), LanguageServer.DidOpenTextDocumentParams(LanguageServer.TextDocumentItem("testdoc", "julia", 0, testtext))), server)
+LanguageServer.process(LanguageServer.JSONRPC.Request{Val{Symbol("textDocument/didOpen")},LanguageServer.DidOpenTextDocumentParams}(0, LanguageServer.DidOpenTextDocumentParams(LanguageServer.TextDocumentItem("testdoc", "julia", 0, testtext))), server)
 
 doc = server.documents[LanguageServer.URI2("testdoc")]
 LanguageServer.parse_all(doc, server)
